+++ src/scriptlets/trusted-set-local-storage-item.js
@@ -29,6 +29,9 @@ import {
  *     - `$currentISODate$` keyword for setting current date in the date time string format,
  *       corresponds to `(new Date).toISOString()` call, e.g '2022-11-08T13:53:19.650Z'
  *
+ * > Keywords listed above can be used as a part of the value as well,
+ * > and more than one keyword can be used in it, e.g `'{"count":1,"firstTime":$now$}'`.
+ *
  * ### Examples
  *
  * 1. Set local storage item
@@ -53,6 +56,12 @@ import {
  *
  *     ```adblock
  *     example.org#%#//scriptlet('trusted-set-local-storage-item', 'player.live.current.play', '$currentDate$')
+ *     ```
+ *
+ * 1. Set item with keywords used as a part of the value
+ *
+ *     ```adblock
+ *     example.org#%#//scriptlet('trusted-set-local-storage-item', 'storage_item', '{"count":1,"firstTime":$now$}')
  *     ```
  *
  * 1. Set item without value